SQL Injection Acknowledgement + Prepared Statement
Using SQL Prepared Statements, trusted code is sent via a code channel, while the untrusted user-provided data are send via a data channel. Therefore, the database clearly knows the boundary between code and data. When it gets data from the data channel, it will not parse the data. Even though an attacker can hide code in data, the code will never be treated as code, so it will never be executed.